Differentitating between output files
