+++ ffgan.py
@@ -14,43 +14,12 @@ import numpy as np
 import scipy.io
 # from tensorflow.examples.tutorials.mnist import input_data
 
-from keras.engine.topology import Layer
 import keras.backend as K
 from keras.models import Sequential
 from keras.layers import Dense, Activation, Flatten, Reshape
-from keras.layers import Conv1D, Conv2DTranspose, UpSampling1D
 from keras.layers import LeakyReLU, Dropout
 from keras.layers import BatchNormalization
-from keras.layers import Lambda
 from keras.optimizers import Adam, RMSprop
-
-class Conv1DTranspose(Layer):
-    """Source: https://github.com/tensorflow/tensorflow/issues/6724#issuecomment-357023018
-    """
-    def __init__(self, filters, kernel_size, strides=1, *args, **kwargs):
-        self._filters = filters
-        self._kernel_size = (1, kernel_size)
-        self._strides = (1, strides)
-        self._args, self._kwargs = args, kwargs
-        super(Conv1DTranspose, self).__init__()
-
-    def build(self, input_shape):
-        print("build", input_shape)
-        self._model = Sequential()
-        self._model.add(Lambda(lambda x: K.expand_dims(x,axis=1), batch_input_shape=input_shape))
-        self._model.add(Conv2DTranspose(self._filters,
-                                        kernel_size=self._kernel_size,
-                                        strides=self._strides,
-                                        *self._args, **self._kwargs))
-        self._model.add(Lambda(lambda x: x[:,0]))
-        self._model.summary()
-        super(Conv1DTranspose, self).build(input_shape)
-
-    def call(self, x):
-        return self._model(x)
-
-    def compute_output_shape(self, input_shape):
-        return self._model.compute_output_shape(input_shape)
 
 class ElapsedTimer(object):
     def __init__(self):
@@ -86,20 +55,17 @@ class DCGAN(object):
         # In: 28 x 28 x 1, depth = 1
         # Out: 14 x 14 x 1, depth=64
         input_shape = (self.img_rows, self.channel)
-        self.D.add(Conv1D(depth*1, 5, strides=2, input_shape=input_shape, padding='same'))
-        self.D.add(LeakyReLU(alpha=0.2))
+        self.D.add(Dense(int(self.img_rows / 2), input_shape=input_shape, activation='tanh'))
+        self.D.add(BatchNormalization(momentum=0.9))
         self.D.add(Dropout(dropout))
-
-        self.D.add(Conv1D(depth*2, 5, strides=2, padding='same'))
-        self.D.add(LeakyReLU(alpha=0.2))
+        self.D.add(Dense(int(self.img_rows / 4), activation='tanh'))
+        self.D.add(BatchNormalization(momentum=0.9))
         self.D.add(Dropout(dropout))
-
-        self.D.add(Conv1D(depth*4, 5, strides=2, padding='same'))
-        self.D.add(LeakyReLU(alpha=0.2))
+        self.D.add(Dense(int(self.img_rows / 8), activation='tanh'))
+        self.D.add(BatchNormalization(momentum=0.9))
         self.D.add(Dropout(dropout))
-
-        self.D.add(Conv1D(depth*8, 5, strides=1, padding='same'))
-        self.D.add(LeakyReLU(alpha=0.2))
+        self.D.add(Dense(int(self.img_rows / 16), activation='tanh'))
+        self.D.add(BatchNormalization(momentum=0.9))
         self.D.add(Dropout(dropout))
 
         # Out: 1-dim probability
@@ -116,43 +82,29 @@ class DCGAN(object):
         dropout = 0.4
         depth = 64 + 64 + 64 + 64  # why is this the depth?
         # dim = 7
-        dim = int(self.img_rows / 4)
+        dim = int(self.img_rows / 16)
         # In: 100
         # Out: dim x dim x depth
-        self.G.add(Dense(dim*depth, input_dim=100))
+        self.G.add(Dense(dim, input_dim=100, activation='sigmoid'))
         self.G.add(BatchNormalization(momentum=0.9))
-        # self.G.add(Activation('relu'))
-        self.G.add(Activation('tanh'))
-        self.G.add(Reshape((dim, depth)))
         self.G.add(Dropout(dropout))
 
-        # for layer in self.G.layers:
-        #     print(type(layer))
-        #     print("input: ", layer.input_shape)
-        #     print("output: ", layer.output_shape)
-
-        # In: dim x dim x depth
-        # Out: 2*dim x 2*dim x depth/2
-        self.G.add(UpSampling1D())
-        self.G.add(Conv1DTranspose(int(depth/2), 5, padding='same'))
+        self.G.add(Dense(dim * 2, input_dim=100, activation='sigmoid'))
         self.G.add(BatchNormalization(momentum=0.9))
-        # self.G.add(Activation('relu'))
-        self.G.add(Activation('tanh'))
-
-        self.G.add(UpSampling1D())
-        self.G.add(Conv1DTranspose(int(depth/4), 5, padding='same'))
+        self.G.add(Dropout(dropout))
+        
+        self.G.add(Dense(dim * 4, input_dim=100, activation='sigmoid'))
         self.G.add(BatchNormalization(momentum=0.9))
-        # self.G.add(Activation('relu'))
-        self.G.add(Activation('tanh'))
+        self.G.add(Dropout(dropout))
 
-        self.G.add(Conv1DTranspose(int(depth/8), 5, padding='same'))
+        self.G.add(Dense(dim * 8, input_dim=100, activation='sigmoid'))
         self.G.add(BatchNormalization(momentum=0.9))
-        # self.G.add(Activation('relu'))
-        self.G.add(Activation('tanh'))
-
-        # Out: 28 x 28 x 1 grayscale image [0.0,1.0] per pix
-        self.G.add(Conv1DTranspose(1, 5, padding='same'))
-        self.G.add(Activation('linear'))
+        self.G.add(Dropout(dropout))
+        
+        self.G.add(Dense(self.img_rows, input_dim=100, activation='linear'))
+        self.G.add(BatchNormalization(momentum=0.9))
+        self.G.add(Dropout(dropout))
+        self.G.add(Reshape((-1, 1)))
         self.G.summary()
 
         return self.G
@@ -178,7 +130,7 @@ class DCGAN(object):
             metrics=['accuracy'])
         return self.AM
 
-class BIO_DCGAN(object):
+class FFGAN(object):
     def __init__(self, dataset):
         self.channel = 1
         self.dataset = dataset
@@ -192,7 +144,6 @@ class BIO_DCGAN(object):
         if self.img_rows != self.x_train.shape[1]:
             self.x_train = np.pad(self.x_train, ((0,0), (0, self.img_rows - self.x_train.shape[1])), 'minimum')
         self.x_train = self.x_train.reshape(self.x_train.shape[0], self.img_rows, 1).astype(np.float32)
-        print(self.x_train.shape)
 
         self.DCGAN = DCGAN(self.img_rows, self.img_cols)
         self.discriminator =  self.DCGAN.discriminator_model()
@@ -224,20 +175,19 @@ class BIO_DCGAN(object):
                     self.plot_images(save2file=True, samples=noise_input.shape[0], noise=noise_input, step=(i+1))
 
     def plot_images(self, save2file=False, fake=True, samples=16, noise=None, step=0):
-        filename = "{}_real.png".format(self.dataset)
+        filename = "{}_{}_real.png".format(self.__class__.__name__, self.dataset)
         if fake:
-            filename="{}_fake.png".format(self.dataset)
+            filename="{}_{}_fake.png".format(self.__class__.__name__, self.dataset)
             if noise is None:
                 noise = np.random.uniform(-1.0, 1.0, size=[samples, 100])
             else:
-                filename = "{}_fake_{}.png".format(self.dataset, step)
+                filename = "{}_{}_fake_{}.png".format(self.__class__.__name__, self.dataset, step)
             images = self.generator.predict(noise)
         else:
             i = np.random.randint(0, self.x_train.shape[0], samples)
             images = self.x_train[i, :, :]
 
         plt.figure(figsize=(2, 4))
-        print("shape of images: ", images.shape)
         # TODO: increase the width of each "image"
         for i in range(images.shape[0]):
             plt.subplot(samples, 1, i+1)
@@ -253,7 +203,7 @@ class BIO_DCGAN(object):
             plt.show()
 
 if __name__ == '__main__':
-    mnist_dcgan = BIO_DCGAN("ALLAML")
+    mnist_dcgan = FFGAN("ALLAML")
     timer = ElapsedTimer()
     mnist_dcgan.train(train_steps=10, batch_size=1, save_interval=10)
     timer.elapsed_time()
